Shuffle some tests around and clean them up.
