+++ test/test_chemistry_cif.py
@@ -15,8 +15,7 @@
 Test cases for reading PDBx/mmCIF data files PdbxReader class -
 
 """
-import sys, unittest, traceback
-import sys, time, os, os.path, shutil
+import sys, unittest, traceback, os
 from utils import get_fn, get_saved_fn, diff_files
 
 from chemistry.pdbx import PdbxReader, PdbxWriter
